Api with validation

+++ controllers/todo.js
@@ -1,4 +1,5 @@
 const Todo = require("../models/todo");
+const { todoSchema, updateSchema } = require("../middleware/joi");
 
 exports.findData = async function (req, res) {
   try {
@@ -15,13 +16,14 @@ exports.findData = async function (req, res) {
 
 exports.findUser = async function (req, res) {
   try {
-    const user = await Todo.findById(req.params.id);
-    console.log(user, "users");
-    res.json({
-      success: true,
-      message: "get details Successfully",
-      data: user,
-    });
+    const data = await Todo.findById(req.params.id);
+    if (!data) throw new Error("id is not found");
+    else
+      return res.json({
+        success: true,
+        message: "get details Successfully",
+        data,
+      });
   } catch (err) {
     res.json({ success: false, message: err.message });
   }
@@ -29,8 +31,8 @@ exports.findUser = async function (req, res) {
 
 exports.addData = async function (req, res) {
   try {
-    let createData = new Todo(req.body);
-    console.log("createData", createData);
+    let value = await todoSchema.validateAsync(req.body);
+    let createData = new Todo(value);
     createData.save();
     res.json({
       success: true,
@@ -43,21 +45,24 @@ exports.addData = async function (req, res) {
 
 exports.updateData = async function (req, res) {
   try {
-    const data = await Todo.findByIdAndUpdate(req.params.id, req.body);
-    console.log(data);
-    res.json({ success: true, message: "todo data is update", data: data });
+    const user = await Todo.findById(req.params.id);
+    if (!user) throw new Error("id is not found");
+    let value = await updateSchema.validateAsync(req.body);
+    const data = await Todo.findByIdAndUpdate(req.params.id, value);
+    res.json({ success: true, message: "todo data is update", data });
   } catch (err) {
-    res.json(err);
     res.json({ success: false, message: err.message });
   }
 };
+
 exports.deleteData = async function (req, res) {
   try {
+    const data = await Todo.findById(req.params.id);
+    if (!data) throw new Error("id is not found");
     const id = await Todo.findByIdAndDelete(req.params.id);
-    console.log(id, "user");
     res
       .status(200)
-      .json({ success: true, message: "id delete id success", data: id });
+      .json({ success: true, message: "id delete id success", id });
   } catch (err) {
     res.json({ success: false, message: err.message });
   }
